Remove KV API usage from `impl::Deleter` class.
Closes #1354.
