+++ HISTORY.md
@@ -60,6 +60,7 @@
 * Removed cast operators of C++ API objects to their underlying C API objects. This helps prevent inadvertent memory issues such as double-frees.
 * Removed ability to set `null` tile extents on dimensions. All dimensions must now have an explicit tile extent.
 * Changed argument `config` in `Array::consolidate()` from a const-ref to a pointer.
+* Removed default includes of `Map` and `MapSchema`. To use the deprecated KV API temporarily, include `<tiledb/map.h>` explicitly.
 
 # TileDB v1.5.0 Release Notes
